cash amount from preferences and some small changes

+++ OilSavings/Model/APDistance.h
@@ -11,5 +11,5 @@
 @interface APDistance : NSObject
 @property (nonatomic) int distance;
 @property (nonatomic, strong) NSString *text;
-- (id) initWithdistance:(NSInteger) d;
+- (id) initWithdistance:(int) d;
 @end
